Dodate CRUD operacije za zaduzenja

+++ client/src/app/services/project.service.ts
@@ -2,6 +2,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from 'src/environments/environment';
 import { CreateProjectDto, UpdateProjectDto } from '../dto/project/project.dto';
+import { getAuthorizationHeader } from '../helper/header.helper';
 
 const BASE_API_URL = environment.api_url+"/project";
 
@@ -16,64 +17,36 @@ export class ProjectService {
 
   getAll(token: string)
   {
-    return this.httpClient.get(BASE_API_URL, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    })
+    return this.httpClient.get(BASE_API_URL, getAuthorizationHeader(token))
   }
 
   getOne(id: number, token: string)
   {
-    return this.httpClient.get(`${BASE_API_URL}/${id}`, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    })
+    return this.httpClient.get(`${BASE_API_URL}/${id}`, getAuthorizationHeader(token))
   }
 
   add(data: {token: string, project: CreateProjectDto})
   {
-    return this.httpClient.post(BASE_API_URL, data.project, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${data.token}`
-      })
-    })
+    return this.httpClient.post(BASE_API_URL, data.project, getAuthorizationHeader(data.token))
   }
 
   update(data: {token: string, id: number, payload: UpdateProjectDto})
   {
-    return this.httpClient.patch(`${BASE_API_URL}/${data.id}`, data.payload, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${data.token}`
-      })
-    })
+    return this.httpClient.patch(`${BASE_API_URL}/${data.id}`, data.payload, getAuthorizationHeader(data.token))
   }
 
   delete(data: {token: string, id: number})
   {
-    return this.httpClient.delete(`${BASE_API_URL}/${data.id}`, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${data.token}`
-      })
-    })
+    return this.httpClient.delete(`${BASE_API_URL}/${data.id}`, getAuthorizationHeader(data.token))
   }
 
   removeTeamFromProject(projectId: number, teamId: number, token: string)
   {
-    return this.httpClient.delete(`${BASE_API_URL}/${projectId}/team/${teamId}`, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    })
+    return this.httpClient.delete(`${BASE_API_URL}/${projectId}/team/${teamId}`, getAuthorizationHeader(token))
   }
 
   addTeamToProject(projectId: number, teamId: number, token: string)
   {
-    return this.httpClient.patch(`${BASE_API_URL}/team`, {project_id: projectId, team_id: teamId}, {
-      headers: new HttpHeaders({
-        'Authorization': `Bearer ${token}`
-      })
-    })
+    return this.httpClient.patch(`${BASE_API_URL}/team`, {project_id: projectId, team_id: teamId}, getAuthorizationHeader(token))
   }
 }
